% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attachment.R
\name{add_attachment}
\alias{add_attachment}
\title{Add a file attachment to an email message}
\usage{
add_attachment(email, file_path, error_on_missing = TRUE)
}
\arguments{
\item{email}{The email message object, as created by the \code{\link[=compose_email]{compose_email()}}
function. The object's class is \code{email_message}.}

\item{file_path}{The path for the file to be attached.}

\item{error_on_missing}{An option to stop with an error if the file cannot be
located on disk. By default this is set to \code{TRUE}.}
}
\description{
This gives us a simple interface for attaching a file to the email object.
When it comes time to send the email through \code{smtp_send()}, all attachments
(specified by individual calls to \code{add_attachment()}) will be faithfully
transmitted along with the message.
}
