% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{add_column_style}
\alias{add_column_style}
\title{Add inline CSS styles to columns in an HTML table object}
\usage{
add_column_style(html_tbl, columns = NULL, property, values)
}
\arguments{
\item{html_tbl}{an HTML table object that is
created using the \code{build_html_table()}
function.}

\item{columns}{an optional vector of column index
numbers that are to be targeted for transformation.
If nothing is provided here then the style will
be applied to all columns.}

\item{property}{the CSS style property that is
to the added.}

\item{values}{values for the CSS style property.
This should be provided as a single-length
character vector, where multiple values for the
same property are separated by a space character.}
}
\value{
an HTML table object.
}
\description{
Modify column styles for a table within an HTML
table object. This is part of an intermediate set
of step in the \code{build_html_table()} ->
\code{add_column_style()} -> \code{emit_html()}
pattern.
}
