% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc_attachment_helpers.R
\name{attachment_helpers}
\alias{attachment_helpers}
\alias{report_rmd}
\alias{generated_files}
\alias{connect_files}
\title{Email Attachment helpers for RStudio Connect}
\usage{
report_rmd()

generated_files(files = NULL)

connect_files(files = NULL)
}
\value{
a list object of with a class of either \code{connect_report},
\code{generated_files}, or \code{connect_files}.
}
\description{
These helper functions are to be used in the \code{\link[=connect_email]{connect_email()}} function's
\code{attachments} arguments (and enclosed in a \code{list()}).
}
