% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_format.R
\name{blastula_email}
\alias{blastula_email}
\title{R Markdown output format for Blastula email messages}
\usage{
blastula_email(section_divs = TRUE, number_sections = FALSE,
  fig_caption = TRUE, smart = TRUE, dev = "png",
  template = "blastula", ...)
}
\description{
R Markdown output format for Blastula email messages
}
