% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_articles.R
\name{block_articles}
\alias{block_articles}
\title{A block of one or two articles with a multicolumn layout}
\usage{
block_articles(...)
}
\arguments{
\item{...}{One or two calls to \code{article_items()}.}
}
\description{
A block of one or two articles with a multicolumn layout
}
