% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_title.R
\name{block_title}
\alias{block_title}
\title{A block with large title text}
\usage{
block_title(title)
}
\arguments{
\item{title}{A single of paragraph of text.}
}
\description{
A block with large title text
}
