% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{blocks}
\alias{blocks}
\title{An enclosure for all HTML block constructors}
\usage{
blocks(...)
}
\arguments{
\item{...}{One or more \code{block_*()} calls.}
}
\description{
An enclosure for all HTML block constructors
}
