% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds_helpers.R
\name{credential_helpers}
\alias{credential_helpers}
\alias{creds_file}
\alias{creds_key}
\alias{creds}
\title{Helpers for supplying SMTP credentials}
\usage{
creds_file(file)

creds_key(key_name)

creds(provider = NULL, user, sender_name = NULL, host = NULL,
  port = NULL, use_ssl = TRUE)
}
\arguments{
\item{file}{When using the \code{\link[=creds_file]{creds_file()}} credential helper, we need to
specify the location of the credential file, and, this is where that is
done. The credential file was ideally generated by the
\code{\link[=create_smtp_creds_file]{create_smtp_creds_file()}} function.}

\item{key_name}{When using the \code{\link[=creds_key]{creds_key()}} credential helper, the name of
the key (in the system key-value store) needs to be given here. This was
either explicitly provided when using the \code{\link[=create_smtp_creds_key]{create_smtp_creds_key()}}
function (with its own \code{key_name} argument), or, it may have been
automatically generated based on the SMTP host name.}

\item{provider}{An optional email provider shortname for autocompleting STMP
configuration details (the \code{host}, \code{port}, \code{use_ssl} options). Options
currently include \code{gmail}, \code{outlook}, and \code{office365}. If nothing is
provided then values for \code{host}, \code{port}, and \code{use_ssl} are expected.}

\item{user}{The username for the email account. Typically, this is the email
address associated with the account.}

\item{sender_name}{An option to specify a sender name. This isn't always
visible to the recipient, however, as some SMTP servers will suppress this.}

\item{host}{Configuration info for the SMTP server. The \code{host}
and \code{port} parameters are the address and port for the SMTP server;
\code{use_ssl} is an option as to whether to use SSL: supply a \code{TRUE} or \code{FALSE}
value (\code{TRUE} is the default value).}

\item{port}{Configuration info for the SMTP server. The \code{host}
and \code{port} parameters are the address and port for the SMTP server;
\code{use_ssl} is an option as to whether to use SSL: supply a \code{TRUE} or \code{FALSE}
value (\code{TRUE} is the default value).}

\item{use_ssl}{Configuration info for the SMTP server. The \code{host}
and \code{port} parameters are the address and port for the SMTP server;
\code{use_ssl} is an option as to whether to use SSL: supply a \code{TRUE} or \code{FALSE}
value (\code{TRUE} is the default value).}
}
\value{
a list object of class \code{location_cells}.
}
\description{
These helper functions, the credential helpers, are used to supply SMTP
configuration and authorization information for the \code{\link[=smtp_send]{smtp_send()}} function.
These helpers (\code{\link[=creds_file]{creds_file()}}, \code{\link[=creds_key]{creds_key()}}, and \code{\link[=creds]{creds()}}) are to be used
expressly with the \code{credentials} argument of \code{\link[=smtp_send]{smtp_send()}}. The
\code{\link[=creds_file]{creds_file()}} credential helper is used to point at a credentials file
stored on disk. We can create that file using the \code{\link[=create_smtp_creds_file]{create_smtp_creds_file()}}
function. The \code{\link[=creds_key]{creds_key()}} credential helper is truly somthing special. It
actually helps one obtain credentials that are stored in the system-wide
key-value store. The equally magical \code{\link[=create_smtp_creds_key]{create_smtp_creds_key()}} function sets
that key. Using \code{\link[=creds]{creds()}} allows for manual specification of SMTP
configuration and authentication within that helper function.
}
