% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_str.R
\name{get_html_str}
\alias{get_html_str}
\title{Get the HTML content of an email message}
\usage{
get_html_str(message)
}
\arguments{
\item{message}{the email message object, as created by the
\code{compose_email()} function. The object's class is \code{email_message}}
}
\value{
a character object containing the email message's HTML content.
}
\description{
Get the HTML content string from an \code{email_message} object as a
single-length character vector.
}
