% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{md}
\alias{md}
\title{Interpret input text as Markdown-formatted text}
\usage{
md(text)
}
\arguments{
\item{text}{The text that is understood to contain Markdown formatting.}
}
\value{
A character object that is tagged for a Markdown-to-HTML
transformation.

a character object of class \code{from_markdown}.
}
\description{
Interpret input text as Markdown-formatted text
}
