% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify.R
\name{minify}
\alias{minify}
\title{Minify an email}
\usage{
minify(email, binary_loc = NULL, minify_opts = NULL, echo = FALSE)
}
\arguments{
\item{email}{The email message object, as created by the \code{\link[=compose_email]{compose_email()}}
function. The object's class is \code{email_message}.}

\item{binary_loc}{An option to supply the location of the \code{minify}
binary file should it not be on the system path or in the working
directory.}

\item{minify_opts}{A list of additional options to pass to the \code{minify} CLI
command. Only the following sensible subset of
\href{https://github.com/tdewolff/minify/tree/master/cmd/minify#usage}{all possible minify options}
is supported, listed with their default values:
\itemize{
\item \code{`html-keep-conditional-comments` = FALSE}: Preserve all IE conditional comments
\item \code{`html-keep-default-attrvals` = FALSE}: Preserve default attribute values
\item \code{`html-keep-document-tags` = FALSE}: Preserve \code{<html>}, \code{<head>} and \code{<body>} tags
\item \code{`html-keep-end-tags` = FALSE}: Preserve all end tags
\item \code{`html-keep-whitespace` = FALSE}: Preserve whitespace characters but still collapse multiple into one
\item \code{`css-decimals` = -1L}: Number of decimals to preserve in CSS numbers,
\code{-1L} means all
\item \code{`svg-decimals` = -1L}: Number of decimals to preserve in SVG numbers,
\code{-1L} means all
\item \code{verbose = FALSE}: Print informative messages about minification details.
}}

\item{echo}{If set to \code{TRUE}, the command to minify the \code{email_message}
object's HTML via \code{minify} will be printed to the console. By default,
this is \code{FALSE}.}
}
\value{
An \code{email_message} object.
}
\description{
Minify a blastula \code{email_message} object.
}
\details{
The \href{https://en.wikipedia.org/wiki/Minification_(programming)}{minification}
relies on the binary
\href{https://github.com/tdewolff/minify/tree/master/cmd/minify}{minify}
which is cross-platform and works on Windows, macOS, Linux and BSD.
Pre-built binaries can be downloaded from
\href{https://github.com/tdewolff/minify/releases}{here}. Alternatively,
instructions to build minify from source are available
\href{https://github.com/tdewolff/minify/tree/master/cmd/minify#installation}{here}.
}
\examples{
\donttest{# Create a simple test email
test_mail <- prepare_test_message()

# Minify the test email
minify(test_mail)

# The command used to minify can be printed
minify(email = test_mail,
       echo = TRUE)

# We can also provide options to the
# underlying minify command
minify(email = test_mail,
       minify_opts = list(`html-keep-conditional-comments` = TRUE,
                          `html-keep-default-attrvals` = TRUE,
                          verbose = TRUE),
       echo = TRUE)
}

}
