% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_email.R
\name{preview_email}
\alias{preview_email}
\title{Preview an email message}
\usage{
preview_email(email)
}
\arguments{
\item{email}{an \code{email_message} object.}
}
\description{
Preview an HTML email in the Viewer before sending it out.
}
