% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blastula_message}
\alias{print.blastula_message}
\title{Print the email object to the Viewer}
\usage{
\method{print}{blastula_message}(x, ...)
}
\arguments{
\item{x}{an \code{email_message} object.}
}
\description{
This function will provide a preview of the email message.
}
\keyword{internal}
