% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_articles.R
\name{print.block_articles}
\alias{print.block_articles}
\title{Print a block of articles}
\usage{
\method{print}{block_articles}(x, ...)
}
\arguments{
\item{x}{an object of class \code{block_articles}.}
}
\description{
This facilitates printing of a block of articles to the Viewer.
}
\keyword{internal}
