% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_social_links.R
\name{print.block_social_links}
\alias{print.block_social_links}
\title{Print a block of social links}
\usage{
\method{print}{block_social_links}(x, ...)
}
\arguments{
\item{x}{an object of class \code{block_social_links}.}
}
\description{
This facilitates printing of a block of social links to the Viewer.
}
\keyword{internal}
