% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_spacer.R
\name{print.block_spacer}
\alias{print.block_spacer}
\title{Print a spacer}
\usage{
\method{print}{block_spacer}(x, ...)
}
\arguments{
\item{x}{an object of class \code{block_spacer}.}
}
\description{
This facilitates printing of the spacer to the Viewer.
}
\keyword{internal}
