% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_text.R
\name{print.block_text}
\alias{print.block_text}
\title{Print a text block}
\usage{
\method{print}{block_text}(x, ...)
}
\arguments{
\item{x}{an object of class \code{block_text}.}
}
\description{
This facilitates printing of the text block to the Viewer.
}
\keyword{internal}
