% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_title.R
\name{print.block_title}
\alias{print.block_title}
\title{Print a title block}
\usage{
\method{print}{block_title}(x, ...)
}
\arguments{
\item{x}{an object of class \code{block_title}.}
}
\description{
This facilitates printing of the title block to the Viewer.
}
\keyword{internal}
