% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_social_links.R
\name{print.social_link}
\alias{print.social_link}
\title{Print a social link component in the console}
\usage{
\method{print}{social_link}(x, ...)
}
\arguments{
\item{x}{an object of class \code{social_link}.}
}
\description{
This facilitates printing of a social link object to the console.
}
\keyword{internal}
