% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email_out_2.R
\name{send_email_out_2}
\alias{send_email_out_2}
\title{Send an email message // Experimental implementation}
\usage{
send_email_out_2(message, subject = NULL, from = NULL, to = NULL,
  cc = NULL, bcc = NULL, attachments = NULL, creds_file = NULL,
  sender = NULL, host = NULL, port = NULL, user = NULL,
  password = NULL, use_ssl = TRUE, use_tls = FALSE, authenticate = TRUE,
  verbose = FALSE, debug = FALSE)
}
\arguments{
\item{message}{the email message object,
as created by the \code{compose_email()}
function. The object's class is
\code{email_message}}

\item{subject}{the subject of the
email.}

\item{from}{the email address of the
sender. This does not have to be
the same email that is associated with
the account actually sending the message.}

\item{to}{a vector of email
addresses.}

\item{cc}{a vector of email
addresses.}

\item{bcc}{a vector of email
addresses.}

\item{attachments}{a vector of paths
to files to be attached to the email.}

\item{creds_file}{an optional path to
an email credentials file. This file
must be created by the
\code{create_email_creds_file}
function.}

\item{sender}{the sender name.}

\item{host}{the email host.}

\item{port}{the port associated with
the email account.}

\item{user}{the username associated
with the email account.}

\item{password}{the password associated
with the email account.}

\item{use_ssl}{a logical value to
indicate whether to use SSL.}

\item{use_tls}{a logical value to
indicate whether to use TLS.}

\item{authenticate}{a logical value to
indicate whether to use authenication.}

\item{verbose}{a logical value indicating
whether verbose messages should be
printed to the console during sending
of email.}

\item{debug}{a logical value to indicate
whether the mail sending statement
should be printed to the console. No
emails are sent when debug is set to
\code{TRUE}.}
}
\description{
Send an email message to one
or more recipients.
}
