% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_email_creds_file.R
\name{set_smtp_credentials}
\alias{set_smtp_credentials}
\title{Store SMTP credentials in the system's key-value store}
\usage{
set_smtp_credentials(key_name = NULL, provider = NULL, user = NULL,
  sender = NULL, host = NULL, port = NULL, use_ssl = TRUE)
}
\arguments{
\item{key_name}{An option to specify an identifier for the stored credentials
in the keyring. If no name is provided, then the provider name will be used
if available.}

\item{provider}{An optional email provider name with which an STMP account is
available. Options currently include \code{gmail}, \code{outlook}, and \code{office365}.
If a \code{provider} is not given, then values for \code{host}, \code{port}, and \code{use_ssl}
will be used.}

\item{user}{The username for the email account.}

\item{sender}{The sender name.}

\item{host}{The \code{host} name.}

\item{port}{The port number.}

\item{use_ssl}{An option as to whether to use SSL; supply a \code{TRUE} or \code{FALSE}
value (\code{TRUE} is the default value).}
}
\description{
Set SMTP access credentials using the \strong{keyring} package.
}
\examples{
\dontrun{
# Store SMTP crendentials using the system's
# secure key-value store
set_stmp_credentials(
  provider = "gmail",
  user = "user_name@gmail.com",
  )
}
}
