% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smtp_send.R
\name{smtp_send}
\alias{smtp_send}
\title{Send an email message through SMTP}
\usage{
smtp_send(email, from, to, subject = NULL, cc = NULL, bcc = NULL,
  credentials = NULL, binary_loc = NULL, echo = FALSE,
  echo_cmd = FALSE, ..., dry_run = FALSE, creds_file = "deprecated")
}
\arguments{
\item{email}{The email message object, as created by the \code{compose_email()}
function. The object's class is \code{email_message}}

\item{from}{The email address of the sender. This does not have to be the
same email that is associated with the account actually sending the
message.}

\item{to}{A vector of email addresses serving as primary recipients for the
message. For secondary recipients, use the \code{cc} and \code{bcc} arguments.}

\item{subject}{The subject of the message, which is usually a brief summary
of the topic of the message.}

\item{cc}{A vector of email addresses for sending the message as a carbon
copy. This list of for those who are to receive a copy of a message
addressed primarily to another. The list of recipients in the CC list is
visible to all other recipients of the message.}

\item{bcc}{A vector of email addresses for sending the message as blind
carbon copies. Any email addresses provided here will receive the message
and these email addresses will be concealed from other recipients
(including others on the BCC list).}

\item{credentials}{One of three credential helper functions must be used
here: (1) \code{\link[=creds_file]{creds_file()}}, (2) \code{\link[=creds_key]{creds_key()}}, or (3) \code{\link[=creds]{creds()}}. The first,
\code{\link[=creds_file]{creds_file()}}, relies on a credentials file stored on disk. Such a file is
created using the \code{\link[=create_smtp_creds_file]{create_smtp_creds_file()}} function. The \code{\link[=creds_key]{creds_key()}}
function is used if credentials are stored in the system-wide key-value
store, through use of the \code{\link[=create_smtp_creds_key]{create_smtp_creds_key()}} function. Using
\code{\link[=creds]{creds()}} allows for a manual specification of SMTP configuration and
credentials within that helper function.}

\item{binary_loc}{An option to supply the location of the \code{mailsend-go}
binary file should it not be on the system path or in the working
directory.}

\item{echo}{An option to print the standard output and error to the screen.}

\item{echo_cmd}{A logical value indicating whether the system command should
be printed to the console during the sending of email.}

\item{...}{The \code{...} is unused and only serves to force the naming of
subsequent argument (\code{dry_run}) and avoid an unintended use.}

\item{dry_run}{Setting \code{dry_run} to \code{TRUE} will return information on the
SMTP sending options. Furthermore, the function will stop short of actually
sending the email message out. By default, however, this is set to \code{FALSE}.}
}
\description{
Send an email message to one or more recipients via an SMTP server.
}
\examples{
\dontrun{
# Prepare a test message and send
# the email out with `smtp_send()`
prepare_test_message() \%>\%
  smtp_send(
    from = "sender@mail.com",
    to = "recipient@mail.com",
    subject = "Mail Subject",
    credentials = creds_file("mail_creds")
  )
}
}
