% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_social_links.R
\name{social_link}
\alias{social_link}
\title{Specify the components of a social link}
\usage{
social_link(service, link, icon = NULL, variant = NULL, alt = NULL)
}
\arguments{
\item{service}{Either the name of a social media service or either of
\code{website}, \code{email}, or \code{rss}.}

\item{link}{The relevant link to content on the \code{service}.}

\item{icon}{A link to an icon for the service. If \code{NULL}, then a link to a
suitable image asset will be automatically created.}

\item{variant}{The variant of the icon to use. Options include \code{bw} (black
and white, the default), \code{color}, \code{dark_gray}, \code{gray}, and \code{light_gray}.}

\item{alt}{Text description of image passed to the \code{alt} attribute inside of
the image (\code{<img>}) tag for use when image loading is disabled and on
screen readers. \code{NULL} default produces blank (\code{""}) alt text.}
}
\description{
The \code{social_link()} function is used exclusively within
\code{block_social_links()} with as many calls as the number of social media links
required.
}
