% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_email.R
\name{suppress_scheduled_email}
\alias{suppress_scheduled_email}
\title{A command to suppress any scheduled emailing in RStudio Connect}
\usage{
suppress_scheduled_email(suppress = TRUE)
}
\arguments{
\item{suppress}{A logical value for whether email suppression should occur
after publication.}
}
\description{
A command to suppress any scheduled emailing in RStudio Connect
}
