% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readable_time.R
\name{add_readable_time}
\alias{add_readable_time}
\title{Create an HTML fragment for a humanized date/time}
\usage{
add_readable_time(time = NULL, use_date = TRUE, use_time = TRUE,
  use_tz = TRUE)
}
\arguments{
\item{time}{The POSIXct time to use, and to make more readable for email
recipients. If a \code{time} is not provided, the current system time will be
used.}

\item{use_date}{A logical value that indicates whether the current date
should be included.}

\item{use_time}{A logical value that indicates whether the current time
should be included.}

\item{use_tz}{A logical value that indicates whether the locale's time zone
should be included.}
}
\value{
A character object that can be placed inside any message component
message wherever the function is called.
}
\description{
Add a nicely-formatted date/time string inside the body of the email with
this helper function. We can provide a POSIXct date-time object or use the
current date/time/tz (based on the caller's locale information at the time of
the call). There are options to specify whether the date, time, and time zone
parts are to be included.
}
