% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_articles.R
\name{article_items}
\alias{article_items}
\title{Specify the components of an article}
\usage{
article_items(image = NULL, title = NULL, content = NULL)
}
\arguments{
\item{image}{An optional URL pointing to an image resource.}

\item{title}{An optional title for the article.}

\item{content}{An optional paragraph of text for the article.}
}
\description{
Specify the components of an article
}
