% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_format.R
\name{attach_email}
\alias{attach_email}
\title{Attach a Blastula email message object to the current render}
\usage{
attach_email(email, preview = TRUE)
}
\description{
Attach a Blastula email message object to the current render
}
