% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_format.R
\name{blastula_email}
\alias{blastula_email}
\title{R Markdown output format for Blastula email messages}
\usage{
blastula_email(section_divs = TRUE, number_sections = FALSE,
  toc = FALSE, toc_depth = 3, toc_float = FALSE, fig_width = 5.35,
  fig_height = 5, fig_retina = 2, fig_caption = TRUE, smart = TRUE,
  dev = "png", self_contained = TRUE, template = "blastula",
  includes = NULL, keep_md = FALSE, md_extensions = NULL, ...)
}
\description{
R Markdown output format for Blastula email messages
}
