% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_format.R
\name{blastula_email}
\alias{blastula_email}
\title{R Markdown output format for Blastula emails}
\usage{
blastula_email(toc = FALSE, toc_depth = 3, toc_float = FALSE,
  number_sections = FALSE, section_divs = TRUE, fig_width = 7,
  fig_height = 5, fig_retina = 2, fig_caption = TRUE, dev = "png",
  df_print = "default", code_folding = c("none", "show", "hide"),
  code_download = FALSE, smart = TRUE, self_contained = TRUE,
  theme = "default", highlight = "default", mathjax = NULL,
  template = "blastula", extra_dependencies = NULL, css = NULL,
  includes = NULL, keep_md = FALSE, lib_dir = NULL,
  md_extensions = NULL, pandoc_args = NULL, ...)
}
\description{
R Markdown output format for Blastula emails
}
