% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_title.R
\name{block_title}
\alias{block_title}
\title{A block with large title text}
\usage{
block_title(...)
}
\arguments{
\item{...}{Paragraphs of title text.}
}
\description{
With \code{block_title()} we can define a title text area and this can be easily
combined with other \code{block_*()} functions. The title will take the entire
width of the block and will resize according to screen width. Like all
\code{block_*()} functions, \code{block_title()} must be placed inside of \code{blocks()}
and the resultant \code{blocks} object can be provided to the \code{body}, \code{header}, or
\code{footer} arguments of \code{compose_email()}.
}
\examples{
# Create a block of two, side-by-side
# articles with two `article()` calls
# inside of `block_articles()`, itself
# placed in `blocks()`; also, include a
# title at the top with `block_title()`
email <-
  compose_email(
    body =
      blocks(
        block_title("Two Cities I Visited Recently"),
        block_articles(
          article(
            image = "https://i.imgur.com/dig0HQ2.jpg",
            title = "Los Angeles",
            content =
              "I want to live in Los Angeles.
              Not the one in Los Angeles.
              No, not the one in South California.
              They got one in South Patagonia."
          ),
          article(
            image = "https://i.imgur.com/RUvqHV8.jpg",
            title = "New York",
            content =
              "Start spreading the news.
              I'm leaving today.
              I want to be a part of it.
              New York, New York."
          )
        )
      )
    )

if (interactive()) email

}
