% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_title.R
\name{block_title}
\alias{block_title}
\title{A block with large title text}
\usage{
block_title(title)
}
\arguments{
\item{title}{A single of paragraph of text.}
}
\description{
With \code{block_text()} we can define a title text area and this can be easily
combined with other \code{block_*()} functions. The title will take the entire
width of the block and will resize according to screen width. Like all
\code{block_*()} functions, \code{block_title()} must be placed in a list object and
that list can only be provided to the \code{blocks} argument of \code{compose_email()}.
}
\examples{
# Create a block of two, side-by-side
# articles with two `article_items()` calls
# inside of `block_articles()`, itself
# placed in a `list()`; also, include a
# title at the top with `block_title()`
compose_email(
  blocks =
    list(
      block_title(
        "Two Cities I Visited Recently"),
      block_articles(
        article_items(
          image = "https://i.imgur.com/dig0HQ2.jpg",
          title = "Los Angeles",
          content =
            "I want to live in Los Angeles. \\\\
            Not the one in Los Angeles. \\\\
            No, not the one in South California. \\\\
            They got one in South Patagonia."
        ),
        article_items(
          image = "https://i.imgur.com/RUvqHV8.jpg",
          title = "New York",
          content =
            "Start spreading the news. \\\\
            I'm leaving today. \\\\
            I want to be a part of it. \\\\
            New York, New York."
        )
      )
    )
  )
}
