% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_email.R
\name{connect_email}
\alias{connect_email}
\title{Include a email message when publishing R Markdown documents in Connect}
\usage{
connect_email(email, subject = NULL, preview = TRUE,
  attachments = NULL)
}
\arguments{
\item{email}{A rendered blastula email. Normally, we'd want to use an
associated .Rmd file with the \code{blastula::blastula_email} R Markdown output
format in the following call: \code{blastula::render_email(input = <email_document>.Rmd)}.}

\item{subject}{An option to specify the the email subject while attaching the
email object.}

\item{preview}{Should the email message display it's own preview window? If
\code{TRUE} (the default), the rendered email message will be shown.}

\item{attachments}{A list of attachments for the Connect email.}
}
\description{
Include a email message when publishing R Markdown documents in Connect
}
