% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_test_message.R
\name{prepare_test_message}
\alias{prepare_test_message}
\title{Prepare a email test message object}
\usage{
prepare_test_message(incl_ggplot = FALSE, incl_image = FALSE)
}
\arguments{
\item{incl_ggplot}{an option to include a ggplot plot within the body of the
test message. This requires that the ggplot2 package is installed. By
default, this is FALSE.}

\item{incl_image}{an option to include a test image within the body of the
test message. By default, this is FALSE.}
}
\description{
Create an email test message object, which is helpful for sending a test
message with the \code{send_email_out()} function.
}
\examples{
\dontrun{
# Create a credentials file to
# send via Gmail (this will be named
# `.bls_smtp_gmail_com`)
create_email_creds_file(
  user = "username@gmail.com",
  password = "*************",
  provider = "gmail",
  sender = "Sender Name")

# Send oneself a test message to
# test these new SMTP settings and also
# to ensure that the message appears
# correctly in the email client
send_email_out(
  message = prepare_test_message(),
  sender = "Sender Name",
  subject = "test 2",
  from = "username@gmail.com",
  to = "username@gmail.com",
  creds_file = ".bls_smtp_gmail_com")
}
}
