% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_articles.R
\name{print.article}
\alias{print.article}
\title{Print the article information in the console}
\usage{
\method{print}{article}(x, ...)
}
\arguments{
\item{x}{an object of class \code{article}.}
}
\description{
This facilitates printing of the article object to the console.
}
\keyword{internal}
