% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_format.R
\name{render_email}
\alias{render_email}
\title{R Markdown render function for the \code{blastula_email} output format}
\usage{
render_email(input, envir = parent.frame(), quiet = TRUE)
}
\arguments{
\item{input}{The input file to be rendered. This should be an R Markdown
document (.rmd) with the output specified as \code{output: blastula::blastula_email}.}

\item{envir}{The environment in which the code chunks are to be evaluated
during knitting (we can use \code{new.env()} to guarantee an empty, new
environment).}

\item{quiet}{An option to suppress printing of the pandoc command line. By
default, this is set to \code{TRUE}.}
}
\description{
R Markdown render function for the \code{blastula_email} output format
}
