% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_by_mailgun.R
\name{send_by_mailgun}
\alias{send_by_mailgun}
\title{Send an email message through the Mailgun API}
\usage{
send_by_mailgun(message, subject = NULL, from, recipient, url, api_key)
}
\arguments{
\item{message}{the email message object,
as created by the \code{compose_email()}
function. The object's class is
\code{email_message}}

\item{subject}{the subject of the
email.}

\item{from}{the email address of the
sender. This does not have to be
the same email that is associated with
the account actually sending the message.}

\item{recipient}{the email address of the
recipient.}

\item{url}{the URL for the sending domain.}

\item{api_key}{the API key registered to
the mailgun service.}
}
\description{
Send an email message via
the Mailgun API. This requires an account
with Mailgun.
}
